requirements.txt
Django==2.2.10
gunicorn==19.7.1
whitenoise==3.3.1
ibmcloudenv
livereload
#pandas==0.24.2
requests
dateparser
mongoengine
pymodm
pymongo==3.10.1
ibm-cos-sdk
bokeh==1.2.0
apscheduler==3.5.1
xlrd==1.2.0
pyfcm
sendgrid==6.3.0
confluent-kafka==1.8.2
pika
openpyxl
djangorestframework